add authentication file.
